remove factoryClass and replace create(factory method) in Figure class so that cis able to check figure type in sub class
