+++ src/main/java/Main.java
@@ -1,5 +1,4 @@
 import domain.Figure;
-import domain.Points;
 import domain.Utils;
 import view.InputView;
 import view.ResultView;
@@ -8,20 +7,18 @@ import java.util.List;
 
 public class Main {
 
-    private static Figure initFigure() {
-        List<String> coordinates;
+    private static List<String> initCoordinates() {
+        System.out.println("Input coordinates. ex. (1,2)-(3,4)-(5,1)");
         try {
-            coordinates = Utils.checkReturnCoordinates(InputView.initCoordinate());
-            return Figure.of(coordinates);
+            return Utils.checkReturnCoordinates(InputView.getCoordinate());
         } catch (RuntimeException e) {
-            return initFigure();
+            return initCoordinates();
         }
     }
 
     public static void main(String[] args) {
         Figure figure;
-        System.out.println("Input coordinates. ex. (1,2)-(3,4)-(5,1)");
-        figure = initFigure();
+        figure = Figure.createFigure(initCoordinates());
         figure.calculate();
         ResultView.printBoard(figure);
         ResultView.printCalculationResult(figure);
