+++ src/main/java/domain/Figure.java
@@ -8,16 +8,6 @@ public class Figure {
 
     double area;
 
-    Figure() { this.points = Points.of(); }
-
-    public static Figure of() {
-        return new Figure();
-    }
-
-    public static Figure of(List<String> coordinates) {
-        return FigureFactory.create(coordinates);
-    }
-
     public boolean hasCoordinateAt(int x, int y) {
         return points.contains(x, y);
     }
@@ -27,5 +17,9 @@ public class Figure {
     public double getArea() {
         return area;
     }
+
+    public static Figure createFigure(List<String> coordinates){
+        return null;
+    };
 
 }
