+++ src/main/java/domain/Line.java
@@ -8,6 +8,17 @@ public class Line extends Figure {
         super.points = Points.initCoordinates(coordinates);
     }
 
+    public static Figure createFigure(List<String> coordinates){
+        if (!isLine(coordinates)){
+            throw new RuntimeException();
+        }
+        return new Line(coordinates);
+    }
+
+    static boolean isLine(List<String> coordinates) {
+        return coordinates.size() == 2;
+    }
+
     @Override
     public void calculate() {
         super.area = super.points.calculateLine();
