+++ src/main/java/view/InputView.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 public class InputView {
 
-    public static String initCoordinate() throws RuntimeException{
+    public static String getCoordinate() throws RuntimeException{
         Scanner scanner = new Scanner(System.in);
         return scanner.nextLine();
     }
